Fixing up file/line reports, level matching
Turns out the entry is reused after logrus fires it. So make a deep
copy of it and manipulate there.

Also, keep the file/line data around as we may be going through these
log entries multiple times.

Fix up going through the cache in one spot.

Also make sure all the places where we should filter on level are doing it.