Rely on timeout to return from non-matching match
That's kinda important.

Also rename a few things to make stuff clearer.

And set the discard/stderr on the actual entry that's used by logrus,
not on our copied entry.

Add some timeouts to the calls in tests around failing matches.

Remove duplicate cacheTop increment: only need to do it once.

Put a mutex around the cache for integrity.

Fix the thing that was modifying the cache.